Fixed Scania constructor and error in Truck

+++ src/labb1/Scania.java
@@ -8,8 +8,8 @@ public class Scania extends Truck{
     private int platformAngle;
 
 // }
-    public Scania(Color color, double enginePower, String modelName, double x, double y, double dirX, double dirY, int maxLoad){
-        super(color, enginePower, modelName, x, y, dirX,  dirY, maxLoad);
+    public Scania(Color color, int enginePower, String modelName, double x, double y, double dirX, double dirY, int maxLoad){
+        super(color, enginePower, modelName, x, y, dirX, dirY, maxLoad);
         this.platformAngle = 0;
     }
